+++ assets/javascripts/kitten/components/cards/crowdfunding-card/widget/index.js
@@ -24,7 +24,7 @@ const StyledTitleAndDescription = styled.div`
     `}
 `
 
-export class WidgetCard extends PureComponent {
+export class CrowdfundingCardWidget extends PureComponent {
   static propTypes = {
     href: PropTypes.string,
     titlesMinHeight: PropTypes.bool,